backend/jcourse_api-master/requirements.txt
Authlib==1.3.2
Django==4.2.16
django-cors-headers==4.6.0
django-filter==24.3
django-import-export==4.2.0
djangorestframework==3.15.2
requests==2.32.2
gunicorn==23.0.0
uvicorn[standard]==0.32.0
psycopg[binary]>=3.2.2
pypinyin==0.53.0
coverage==7.6.4
django-debug-toolbar==4.4.6
redis==5.2.0
python-dotenv==1.0.1
huey==2.5.2
jieba==0.42.1
packaging==24.1
qiniu==7.15.0
Pillow==11.0.0
tencentcloud-sdk-python==3.0.1253
